+ Merge v0.15.0.2 tag
+ Now ./build.sh will produce bitcoin-gold-* instead of bgold-*
+ Fix test according to https://github.com/wilsonmeier/BTCGPU/commit/9fb670da82cdf40beeb23c64b73ae990e03607e7

+++ src/txdb.cpp
@@ -448,7 +448,7 @@ bool CBlockTreeDB::LoadBlockIndexGuts(const Consensus::Params& consensusParams, 
                 pindexNew->nTx            = diskindex.nTx;
 
                 // TODO(h4x3rotab): Check Equihash solution? Not sure why Zcash doesn't do it here.
-                bool postfork = pindexNew->nHeight >= (uint32_t)consensusParams.BTGHeight;
+                bool postfork = pindexNew->nHeight >= consensusParams.BTGHeight;
                 if (!CheckProofOfWork(pindexNew->GetBlockHash(), pindexNew->nBits, postfork, consensusParams))
                     return error("%s: CheckProofOfWork failed: %s", __func__, pindexNew->ToString());
